# server/openapi.yaml
openapi: 3.0.3
info:
  title: MyGamesAnywhere Server API
  description: Public HTTP API for the MyGamesAnywhere server and web interface.
  version: 1.0.0
paths:
...
  /api/device-transfers/archive:
    get:
      responses:
        400:
          description: Error response
        500:
          description: Error response
    head:
      responses:
        400:
          description: Error response
        500:
          description: Error response
...